Iron plant uat merge master

+++ backend/de.metas.ui.web.base/src/main/java/de/metas/ui/web/kpi/data/KPIDataValue.java
@@ -134,7 +134,7 @@ public class KPIDataValue
 					}
 					else if (value instanceof Number)
 					{
-						final BigDecimal bd = BigDecimal.valueOf(((Number)value).intValue());
+						final BigDecimal bd = BigDecimal.valueOf(((Number)value).doubleValue());
 						return roundToPrecision(bd);
 					}
 					else
